Think I'm gonna switch to a hardcoded approach to Kratos, instead using Baritone, and writing a hard-coded algorithm to decide which actions to take based on current health, health of enemy nearest wood block, nearest iron ore, etc. Mineflayer unfortunately simply isn't stable enough. The physics bugs out sometimes, even on 1.17, which is meant to be stable; it will often fail to collect the block is has been set to collect, simply leaving the block on the ground. And even if the bot were able to become 100% stable and reliable, it would simply carry out the functionality of baritone.
Further, with the AI, even though I'd really like to use it, due to the high-level approach of the action commands which will be given (create wooden axe, craft iron armour, etc), it's hard to imagine the AI producing a solution policy which isn't something which would be able to be hardcoded. If the inputs were more low-level (look left, look right, dig, craft block, place block, etc), AI would certainly be the best approach. However, while it would eventually outmatch Baritone (if the stability of the lower-level commands held up, which they certainly appear to), it would take a lot of effort to get the AI to be at the same level of Baritone, while one can simply use Baritone to have that high level of performance

For now, it appears the best approach is to use a hard-coded approach using Baritone, and have this as version 1 of Kratos. Ideally, if people liked Kratos, and one were able to live off the Kratos subscriptions, you would then be able to use the extra time to aim to develop a PPO-based Kratos, using Mineflayer, potentially also investing money into getting devs on the discord to stabilise the Mineflayer bot (the physics, collectBlock, etc)

Overall, the future looks bright for Kratos! There will certainly be challenges and difficulties with implementing Kratos with Baritone, using the Java code, and learning how to code with Baritone, but so far the challenges look surmountable. However, there will likely be many challenges to overcome to be able to have Kratos 1 working